wine-ml/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.13.1
matplotlib==3.9.0
plotly==5.22.0
category-encoders==2.6.3
lightgbm==4.3.0
xgboost==2.1.1
eli5==0.13.0
shap==0.45.1
joblib==1.4.2
openpyxl==3.1.5
python-dotenv==1.0.1
